Content and template installation
A lot of changes with this one, I was hesitant to make a commit for some reason so we have an updated about me, project list, reviews, and contact page. A lot of design choices to make with all this new content so we will see where it takes me.